+++ js/data.js
@@ -1,0 +1,19 @@
+const DESCRIPTION_ADVERB = [
+  'Невероятно',
+  'Восхитительно',
+  'Классно',
+  'Замечательно',
+  'Необычно',
+  'Странно'
+];
+
+const DESCRIPTION_VERB = [
+  'посидели',
+  'отдохнули',
+  'встретились',
+  'погуляли',
+  'поиграли',
+  'покатались',
+];
+
+export {DESCRIPTION_VERB, DESCRIPTION_ADVERB};
