+++ package.json
@@ -11,6 +11,10 @@
     "url": "https://github.com/htmlacademy-javascript/1236243-kekstagram-simple-27/issues"
   },
   "devDependencies": {
+    "@babel/cli": "^7.19.3",
+    "@babel/core": "^7.19.3",
+    "@babel/node": "^7.19.1",
+    "@babel/preset-env": "^7.19.4",
     "browser-sync": "2.27.10",
     "eslint": "^8.15.0",
     "eslint-config-htmlacademy": "6.0.0"
